[Improve]Header,Footerの単一コンポーネント化(#7)
