+++ app/javascript/components/layout/TheHeader.vue
@@ -1,0 +1,23 @@
+<template>
+  <v-app-bar
+    color="grey darken-3"
+    style="height: 70px"
+    app
+  >
+    <v-img 
+      max-height="200"
+      max-width="200"
+      src="../assets/boitore.png"
+    />
+  </v-app-bar>
+</template>
+
+<script>
+export default {
+  name: "TheHeader"
+}
+</script>
+
+<style scoped>
+
+</style>
