+++ app/javascript/packs/application.js
@@ -4,9 +4,9 @@
 // that code so it'll be compiled.
 
 import Rails from '@rails/ujs';
-import Turbolinks from 'turbolinks';
+// import Turbolinks from 'turbolinks';
 import * as ActiveStorage from '@rails/activestorage';
-import 'channels';
+import 'router';
 
 Rails.start();
 Turbolinks.start();
