+++ app/views/top/index.html.erb
@@ -1,4 +1,2 @@
  <%= javascript_pack_tag 'hello_vue' %>
  <%= stylesheet_pack_tag 'hello_vue' %>
- 
-<div id="app"></div>
